Fix 500 on channel list when Dispatcharr uses fractional channel numbers
GET /api/channels crashed with a Pydantic ValidationError whenever a synced
channel had a fractional channel_number (e.g. Dispatcharr subchannel 5.1).
Sync itself succeeded and wrote all rows, but the dashboard's follow-up
fetch to display them 500'd, leaving Total Channels stuck at 0 and Last
Synced: Never with no indication why (surfaced only via the sync-error
banner as "Request failed: 500").

Widen CachedChannel.number (models.py) and ChannelResponse.number
(channels.py) from int to float to match Dispatcharr's actual channel
numbering. No DB migration needed - SQLite has no strict column typing.

+++ backend/channels.py
@@ -20,7 +20,7 @@ router = APIRouter(prefix="/api/channels", tags=["channels"])
 class ChannelResponse(BaseModel):
     id: int
     name: str
-    number: int | None
+    number: float | None
     group_name: str | None
     current_logo_url: str | None
     cache_logo_url: str | None
